+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 ### Added
- - Nothing yet
+ - Test suite to simulate running this addon inside the context of a consuming Ember app using [ember-cli-addon-tests](https://github.com/tomdale/ember-cli-addon-tests)
 
 
 ## [v0.8.0] - 2016-11-11